requirements.txt
opencv-python
face_recognition
imutils
pyobjc-framework-Quartz